extract useAlwaysValidPage from client list and event list and put in own file.
