+++ frontend/client-list/client-list.component.tsx
@@ -6,6 +6,7 @@ import ClientsTable from "./table/clients-table.component";
 import easyFetch from "../util/easy-fetch";
 import { SearchParseValues } from "../util/list-search/search-dsl.helpers";
 import queryString from "query-string";
+import { useAlwaysValidPage } from "../util/use-always-valid-page";
 
 export default function ClientList(props: ClientListProps) {
   const [selectedClients, setSelectedClients] = React.useState<SelectedClients>(
@@ -20,7 +21,7 @@ export default function ClientList(props: ClientListProps) {
     getInitialState
   );
 
-  useAlwaysValidPage(apiState, dispatchApiState);
+  useAlwaysValidPage(apiState, ApiStateStatus, ActionTypes, dispatchApiState);
   useClientsApi(apiState, dispatchApiState);
   useFrontendUrlParams(apiState, dispatchApiState);
 
@@ -168,43 +169,6 @@ function useClientsApi(apiState, dispatchApiState) {
         });
 
       return () => abortController.abort();
-    }
-  }, [apiState]);
-}
-
-function useAlwaysValidPage(apiState, dispatchApiState) {
-  React.useEffect(() => {
-    if (apiState.status !== ApiStateStatus.fetched) {
-      // wait for data first
-      return;
-    }
-
-    const lastPage = Math.ceil(
-      apiState.apiData.pagination.numClients /
-        apiState.apiData.pagination.pageSize
-    );
-
-    let newPage;
-
-    if (
-      typeof apiState.page !== "number" ||
-      isNaN(apiState.page) ||
-      isNaN(lastPage)
-    ) {
-      newPage = 1;
-    } else if (apiState.page <= 0) {
-      newPage = 1;
-    } else if (lastPage === 0) {
-      newPage = 1;
-    } else if (apiState.page > lastPage) {
-      newPage = lastPage;
-    }
-
-    if (newPage && newPage !== apiState.page) {
-      dispatchApiState({
-        type: ActionTypes.newPage,
-        page: newPage,
-      });
     }
   }, [apiState]);
 }
